Updated MainCode file
Had an error where the database couldn't be found when trying to save to it, and figured out that it was because the create_tables() method wasn't run at any point. Since this method creates the database, it should only be run once (on the first time) so that any existing saves aren't deleted when starting the game up again.